backend/requirements.txt
gradio
nltk
sumy
PyPDF2
python-docx
numpy
lxml_html_clean